+++ CHANGELOG.md
@@ -10,7 +10,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 [Unreleased]: https://github.com/LoicGrobol/zeldarose/compare/v0.6.0...HEAD
 
-## Added
+### Added
 
 - The `--tf32-mode` option allows to select the level of NVidia Ampère matmul otpimisations.
 - The `--seed` option allows to fix a random seed.